jira/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
websockets>=14.0
